Don't use uint32 - we don't really need them for this usage. Saves 96 bytes.

+++ parse.c
@@ -21,7 +21,7 @@
 
 BOOL memstrcmp(const uint8_t *start, const uint8_t *end, const char *str)
 {
-    uint32_t len = strlen(str);
+    uint16_t len = strlen(str);
 
     if (len != end-start)
         return FALSE;
@@ -42,7 +42,7 @@ static int8_t digit_to_int(int8_t ch)
 }
 
 // parse binary, hex and decimal numbers
-BOOL parse_number(const uint8_t *str, uint8_t len, uint32_t *result)
+BOOL parse_number(const uint8_t *str, uint8_t len, uint16_t *result)
 {
     uint8_t base = 10;
     uint8_t i;
@@ -123,7 +123,7 @@ BOOL tok(uint8_t const **tok_start, uint8_t const **tok_end, const uint8_t *line
 }
 
 // advance tok_start/tok_end to next token and read number value into i
-BOOL tok_num(const uint8_t **tok_start, const uint8_t **tok_end, const uint8_t *line_end, uint32_t *i)
+BOOL tok_num(const uint8_t **tok_start, const uint8_t **tok_end, const uint8_t *line_end, uint16_t *i)
 {
     if (tok(tok_start, tok_end, line_end))
     {
